[BACKLOG-2540] - AMD Conversion - Splitting ccc_wrapper into multiple files * `VizTypeRegistry` now supports async factories, returning Promises. * Added "common-ui/vizapi/ccc/wrapper/charts/all.js" as a way to load/reference all existing chart implementations (important for bundling) * Made the `vizTypeProvider` leaner by removing the `def.js` dependency (`def.js` is only incurred when you actually wanna show CCC vizs) * Removed ccc wrapper's "IE event bubbling shield" * A bit refactoring of ccc wrapper label configuration functions * A bit formatting of ccc wrapper code
