+++ package-res/resources/web/vizapi/ccc/wrapper/charts/BoxplotChart.js
@@ -1,0 +1,52 @@
+/*!
+* Copyright 2010 - 2015 Pentaho Corporation.  All rights reserved.
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+* http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
+define([
+    "./AbstractCategoricalContinuousChart"
+], function(AbstractCategoricalContinuousChart) {
+
+    return AbstractCategoricalContinuousChart.extend({
+        methods: {
+            _cccClass: 'BoxplotChart',
+
+            _rolesToCccDimensionsMap: {
+                'columns':     null,
+                //'rows':      'category',
+                'multi':       'multiChart',
+                'measures':    'median',
+                'percentil25': 'percentil25',
+                'percentil75': 'percentil75',
+                'percentil5':  'percentil5',
+                'percentil95': 'percentil95'
+            },
+
+            _options: {
+                extensionPoints: {
+                    boxRuleWhisker_strokeDasharray: '- '
+                }
+            },
+
+            _readData: function() {
+                // The boxplot data passes-trough, as is.
+                this.base();
+
+                // In CCC, it is read as a custom format (more relational-like)
+                // Where categoriesCount is the number of "category" dimensions,
+                // not including multichart columns...
+                this.options.dataOptions.categoriesCount = this.axes.row.gemsByRole.rows.length;
+            }
+        }
+    });
+});
